валидация часть 2 и рефакторинг папок
