+++ js/form/form.js
@@ -1,15 +1,13 @@
-import { setDisabledElements } from '../js/util.js';
-import { HOUSE_MIN_PRICE_VALUE, HOUSE_MAX_PRICE_VALUE, RoomsCapacityMap, ERROR_VALIDATION_TEXT } from '../js/constants.js';
+import { HOUSE_MIN_PRICE_VALUE, HOUSE_MAX_PRICE_VALUE, RoomsCapacityMap, ERROR_VALIDATION_TEXT } from './form-constants.js';
 
-const mapFiltersForm = document.querySelector('.map__filters');
-const mapFiltersFormElements  = mapFiltersForm.querySelectorAll('.map__filter');
 const announcementForm = document.querySelector('.ad-form');
-const announcementFormElements = announcementForm.querySelectorAll('.ad-form__element');
 const titleField = announcementForm.querySelector('#title');
 const houseTypeField = announcementForm.querySelector('#type');
 const priceField = announcementForm.querySelector('#price');
 const roomNumberField = announcementForm.querySelector('#room_number');
 const capacityField = announcementForm.querySelector('#capacity');
+const timeIn = announcementForm.querySelector('#timein');
+const timeOut = announcementForm.querySelector('#timeout');
 
 const pristine = new Pristine(announcementForm, {
   classTo: 'ad-form__element',
@@ -19,20 +17,6 @@ const pristine = new Pristine(announcementForm, {
   errorTextTag: 'span',
   errorTextClass: 'form__error'
 });
-
-const inactiveState = () => {
-  announcementForm.classList.add('ad-form--disabled');
-  mapFiltersForm.classList.add('map__filters--disabled');
-  setDisabledElements(announcementFormElements, true);
-  setDisabledElements(mapFiltersFormElements, true);
-};
-
-const activeState = () => {
-  announcementForm.classList.remove('ad-form--disabled');
-  mapFiltersForm.classList.remove('map__filters--disabled');
-  setDisabledElements(announcementFormElements, false);
-  setDisabledElements(mapFiltersFormElements, false);
-};
 
 const validateTitle = (value) => value.length >= 30 && value.length <= 100;
 const validateMinPrice = (value) => value >= HOUSE_MIN_PRICE_VALUE[houseTypeField.value];
@@ -48,6 +32,16 @@ const onPriceChange = () => {
   pristine.validate(priceField);
 };
 
+const validateTimeOut = () => {
+  timeOut.value = timeIn.value;
+  pristine.validate(timeIn);
+};
+
+const validateTimeIn = () => {
+  timeIn.value = timeOut.value;
+  pristine.validate(timeOut);
+};
+
 pristine.addValidator(titleField, validateTitle, ERROR_VALIDATION_TEXT.TITLE);
 pristine.addValidator(priceField, validateMinPrice, getMinPriceErrorMessage);
 pristine.addValidator(priceField, validateMaxPrice, getMaxPriceErrorMessage);
@@ -57,10 +51,10 @@ houseTypeField.addEventListener('change', onPriceChange);
 roomNumberField.addEventListener('change', () => {
   pristine.validate(capacityField);
 });
+timeIn.addEventListener('change', validateTimeOut);
+timeOut.addEventListener('change', validateTimeIn);
 
 announcementForm.addEventListener('submit', (event) => {
   event.preventDefault();
   pristine.validate();
 });
-
-export { inactiveState, activeState };
